langchain-templates/langchain-simple-rag/pyproject.toml
[project]
name = "langchain-simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-chroma>=0.2.1",
    "langchain-community>=0.3.16",
    "langchain-openai>=0.3.3",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
]
